fix(workshop): address session persistence review
